Redesign template configuration
This commit introduces a new module "email". It sends out emails based
on a complete template including the headers, with automatic selection
of language variants, and filling in of arbitary parameters. A few minor
changes are done in other modules for the usage of that new module.

The existing application configuration and templates are migrated to
that new format. The application configuration, now much more
straightforward than before, is changed from a complex object structure
to a simple tree of dictionaries, and a lot of code which has now become
obsolete is removed.